bak/pyproject.toml
[project]
name = "neo4j-azure-ai-workshop"
version = "0.1.0"
description = "Neo4j and Azure Lab - GraphRAG workshop with Microsoft Foundry"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-ai-projects>=1.0.0",
    "azure-ai-inference>=1.0.0b9",
    "python-dotenv",
    "pydantic>=2.0.0,<2.13.0a0",
    "pydantic-settings>=2.13.1",
    "neo4j>=6.1.0",
    "neo4j-graphrag>=1.13.1",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "ipykernel>=7.2.0",
    "nbstripout>=0.9.1",
    "agent-framework-core>=1.0.0rc2",
    "agent-framework-azure-ai>=1.0.0rc2",
    "agent-framework-neo4j>=0.5.4",
    "neo4j-agent-memory[microsoft-agent]>=0.0.4",
    "agent-framework>=1.0.0rc2",
    "spacy>=3.7.0,<4.0.0",
    "en-core-web-sm",
    "rapidfuzz>=3.0.0",
    "nest-asyncio>=1.6.0",
]

[tool.uv]
package = false

# The neo4j-agent-memory[microsoft-agent] extra depends on the `agent-framework`
# meta-package, which depends on agent-framework-core[all] — pulling in every
# provider (Claude, Bedrock, Anthropic, etc.) plus ~3GB of dependencies (nvidia,
# triton, claude_agent_sdk). Replace it with a local shim that only depends on
# agent-framework-core (without [all]).
[tool.uv.sources]
agent-framework = { path = "agent-framework-shim" }
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
